tested auth services

+++ src/lib/auth/index.js
@@ -24,7 +24,7 @@ const register = async ({
     email,
     password,
     role = "customer",
-    account = "",
+    account,
 }) => {
     const hasUser = await userService.isUserExist(email);
     if (hasUser) {
@@ -47,7 +47,8 @@ const register = async ({
  * @param {string} password - password of a user
  * @returns {string} accessToken
  */
-const login = async ({ email, password }) => {
+const login = async ({ email, password, secret }) => {
+    if (!email || !password) throw badRequest();
     const user = await userService.findUserByEmail(email);
     if (!user) {
         throw badRequest("Invalid credentials");
@@ -58,7 +59,7 @@ const login = async ({ email, password }) => {
         throw badRequest("Invalid credentials");
     }
 
-    const accessToken = await createAccessToken({ email, password });
+    const accessToken = await createAccessToken({ email, password, secret });
     return accessToken;
 };
 
@@ -66,7 +67,7 @@ const login = async ({ email, password }) => {
  * @param {string} email - email of a user
  * @returns {string} accessToken
  */
-const createAccessToken = async ({ email }) => {
+const createAccessToken = async ({ email, secret }) => {
     const user = await userService.findUserByEmail(email);
 
     if (!user) {
@@ -80,14 +81,14 @@ const createAccessToken = async ({ email }) => {
         role: user.role,
     };
 
-    return generateToken({ payload });
+    return generateToken({ payload, secret });
 };
 
 /** - create a refresh token
  * @param {string} email - email of a user
  * @returns {string} refresh token
  */
-const createRefreshToken = async ({ email }) => {
+const createRefreshToken = async ({ email, secret }) => {
     const user = await userService.findUserByEmail(email);
     if (!user) {
         throw badRequest("Invalid credentials");
@@ -98,7 +99,7 @@ const createRefreshToken = async ({ email }) => {
         email: user.email,
         // role: user.role,
     };
-    const token = generateToken({ payload, expiresIn: "7d" });
+    const token = generateToken({ payload, expiresIn: "7d", secret });
     const refreshToken = new Refresh({ email, token });
     await refreshToken.save();
 
@@ -109,20 +110,22 @@ const createRefreshToken = async ({ email }) => {
  * @param {string} refresh_token - refresh_token of a user
  * @returns {object} - decoded object of a refresh token
  */
-const verifyRefreshToken = (refresh_token) => {
-    const decoded = tokenService.verifyToken(refresh_token);
+const verifyRefreshToken = (refresh_token, secret) => {
+    if (!refresh_token) throw badRequest();
+    const decoded = tokenService.verifyToken(refresh_token, secret);
     return decoded;
 };
 
 /** - delete a refresh token
  * @param {string} token - refresh token of a user
  */
-const removeRefreshToken = async ({ token = "" }) => {
+const removeRefreshToken = async (token) => {
+    if (!token) throw badRequest();
     const refresh = await Refresh.find({ token: token });
     if (refresh.length < 1) {
         throw notFound("Refresh token does not exist");
     }
-    await Refresh.findOneAndDelete({ token: token });
+    return Refresh.findOneAndDelete({ token: token });
 };
 
 /** -  check expiration of token
@@ -138,6 +141,7 @@ const isExpiredToken = (token) => {
  * @return {string} refesh token
  */
 const findRefreshToken = async (email) => {
+    if (!email) throw badRequest();
     const refresh = await Refresh.findOne({ email: email });
 
     if (refresh.length < 1) {
